requirements.txt
django>=5.2
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
django-filter>=25.1
django-cors-headers>=4.7.0
django-environ>=0.12.0
django-axes>=7.0.2
channels>=4.2.2
daphne>=4.1.2
daphne[http2,tls]>=4.1.2
twisted[tls,http2]
channels-redis>=4.2.1
whitenoise>=6.9.0
pytest-django>=4.11.1
factory-boy>=3.3.3          
bcrypt>=4.3.0   
pillow>=11.2.1
pyyaml>=6.0.2